RYSK-26: Hoist component to shared.
