+++ packages/front-end/src/components/vault/VaultWithdrawBalanceTooltip.tsx
@@ -2,7 +2,7 @@ import { DHV_NAME, ZERO_UINT_256 } from "../../config/constants";
 import { useUserPosition } from "../../hooks/useUserPosition";
 import { Currency } from "../../types";
 import { BigNumberDisplay } from "../BigNumberDisplay";
-import { RyskTooltip } from "../RyskTooltip";
+import { RyskTooltip } from "../shared/RyskToolTip";
 
 export const VaultWithdrawBalanceTooltip = () => {
   const { positionBreakdown } = useUserPosition();
